Implemented coding standrads
